Increase visibility for paypal information
Rearranges and increases the visibility of the Paypal information in regards to receiving money.

Previously, with the information present at the end of the section, it was easy to get visually lost below the fold. Bringing it above the screenshot and using a hint to visually distinguish it helps make sure it is seen by users.